LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select-in-page/customizable-select-in-page-mouse-behavior.tentative-expected.txt



FAIL mouse: input behavior for base appearance <select multiple> assert_equals: There should be two selected options after second click. expected 2 but got 1
FAIL mouse: input behavior for base appearance <select size=4> assert_equals: There should be no selected options after third click. expected 0 but got 1
FAIL touch: input behavior for base appearance <select multiple> assert_equals: There should be one selected option after first click. expected 1 but got 0
FAIL touch: input behavior for base appearance <select size=4> assert_equals: There should be one selected option after first click. expected 1 but got 0
